Make moveCreature emit creature move up and down
It was not working as expected because the statement
oldPos.isInRange(newPos, 1, 1, 0);
was never true.

This commit allows the engine to save data because it will
only send the tiles that are not on the screen.
The above statement is valid for tiles on the same floor or
one floor up and down.